backend/requirements.txt
Flask
Flask-Cors
Flask-MySQLdb
PyJWT
bcrypt
opencv-python
